Refactor pay profiles
