+++ components/payprofile/default.htm
@@ -9,7 +9,7 @@
                 </button>
             </p>
 
-            <input type="hidden" name="redirect" value="{{ returnPage|page }}" />
+            <input type="hidden" name="redirect" value="true" />
          {{ form_close() }}
     {% endif %}
